Splits up files for readability
