+++ OBSCounter/version.h
@@ -2,7 +2,7 @@
 #define VERSION_MAJOR 3
 #define VERSION_MINOR 0
 #define VERSION_PATCH 0
-#define VERSION_BUILD 34
+#define VERSION_BUILD 37
 
 #define stringify(a) stringify_(a)
 #define stringify_(a) #a
